requirements.txt
Django==1.9.4
Markdown==2.6.8
PyYAML==3.11
Pygments==2.1.3
argparse==1.2.1
bleach==1.5.0
blessings==1.6
bpython==0.15
configparser==3.5.0b2
curtsies==0.2.6
dj-database-url==0.4.0
dj-static==0.0.6
django-allauth==0.25.2
django-braces==1.8.1
django-cors-headers==1.1.0
django-extensions==1.6.1
django-oauth-toolkit==0.10.0
django-rest-swagger==0.3.5
django-reversion==1.10.1
django-sslify==0.2.7
django-toolbelt==0.0.1
djangorestframework==3.3.3
env-tools==2.0.0
greenlet==0.4.9
gunicorn==19.4.5
html5lib==0.9999999
myvariant==0.3.1
oauthlib==1.0.3
psycopg2==2.6.1
python-openid==2.2.5
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
static3==0.6.1
tini==3.0.1
wcwidth==0.1.6
whitenoise==2.0.6
wsgiref==0.1.2
